# apps/backend/src/main/java/com/vencentdev/backend/todo/dto/TodoCreateRequest.java
package com.vencentdev.backend.todo.dto;

import com.vencentdev.backend.todo.entity.TodoStatus;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import java.time.LocalDate;

public record TodoCreateRequest(
    @NotBlank @Size(max = 200) String title,
    @Size(max = 5000) String description,
    @NotNull TodoStatus status,
    LocalDate dueDate) {}