# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend API",
    "version": "0.0.1"
  },
  "paths": {
    "/api/v1/admin/ping": {
      "get": {
        "operationId": "adminPing",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "authMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "updateCurrentUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/todos": {
      "get": {
        "operationId": "listTodos",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TodoStatus"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseTodoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "createTodo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/todos/{id}": {
      "get": {
        "operationId": "getTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "updateTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "KycStatus": {
        "type": "string",
        "enum": ["NONE", "PENDING", "VERIFIED", "REJECTED"]
      },
      "PageResponseTodoResponse": {
        "type": "object",
        "required": ["content", "page", "size", "totalElements", "totalPages"],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodoResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Role": {
        "type": "string",
        "enum": ["USER", "ADMIN"]
      },
      "TodoCreateRequest": {
        "type": "object",
        "required": ["title", "status"],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          },
          "description": {
            "type": ["string", "null"],
            "maxLength": 5000
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          },
          "dueDate": {
            "type": ["string", "null"],
            "format": "date"
          }
        }
      },
      "TodoResponse": {
        "type": "object",
        "required": ["id", "title", "status", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": ["string", "null"]
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          },
          "dueDate": {
            "type": ["string", "null"],
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TodoStatus": {
        "type": "string",
        "enum": ["TODO", "IN_PROGRESS", "DONE"]
      },
      "TodoUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": ["string", "null"],
            "maxLength": 200
          },
          "description": {
            "type": ["string", "null"],
            "maxLength": 5000
          },
          "status": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TodoStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "dueDate": {
            "type": ["string", "null"],
            "format": "date"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": ["id", "email", "role", "userType", "kycStatus", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": ["string", "null"]
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          },
          "kycStatus": {
            "$ref": "#/components/schemas/KycStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserType": {
        "type": "string",
        "enum": ["INDIVIDUAL", "ORGANIZATION"]
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": ["string", "null"],
            "format": "email"
          },
          "displayName": {
            "type": ["string", "null"],
            "maxLength": 120
          }
        }
      }
    }
  }
}